src/requirements.txt
# Function dependencies, for example:
# package>=version

fsspec
gcsfs
pandas
numpy
google.cloud
google-cloud-storage 
